Create Centauri_Prime.cpp
Passed the tests.

This program is converting the last letter of the input string to lower case letters and checking if it's a vowel or 'y'.
If it's a vowel the kingdom ruler is alice. If it's y nobody is the ruler. It's a consonant if it's neither y nor a vowel. In this case the ruler is bob.